# perf(python): 2-3x speedup in creating literals/Series of type `Date`