I was refactoring and spotted a performance improvement opportunity.

Instead of converting to datetime and going through that path, there is now a dedicated path for dates.

* Internal conversion util `_date_to_pl_date` :  1.43 µs -> 23.4 ns
* Creating a literal expression `lit` :  23.4 µs -> 10.8 µs
* Creating a Series from a list of a million dates : 2.97s -> 1.14s